# frontend/src/api/generated/openapi.json
{
  "components": {
    "schemas": {
      "AnalysisObservationErrorResponse": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "enum": [
              "invalid_fen",
              "analysis_unavailable",
              "unexpected_failure"
            ],
            "title": "Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "AnalysisObservationErrorResponse",
        "type": "object"
      },
      "AnalysisObservationLineResponse": {
        "additionalProperties": false,
        "properties": {
          "depth": {
            "minimum": 0.0,
            "title": "Depth",
            "type": "integer"
          },
          "pv_uci": {
            "items": {
              "type": "string"
            },
            "title": "Pv Uci",
            "type": "array"
          },
          "rank": {
            "minimum": 1.0,
            "title": "Rank",
            "type": "integer"
          },
          "score_kind": {
            "enum": [
              "cp",
              "mate"
            ],
            "title": "Score Kind",
            "type": "string"
          },
          "score_value": {
            "title": "Score Value",
            "type": "integer"
          },
          "wdl_draws": {
            "minimum": 0.0,
            "title": "Wdl Draws",
            "type": "integer"
          },
          "wdl_losses": {
            "minimum": 0.0,
            "title": "Wdl Losses",
            "type": "integer"
          },
          "wdl_wins": {
            "minimum": 0.0,
            "title": "Wdl Wins",
            "type": "integer"
          }
        },
        "required": [
          "rank",
          "score_kind",
          "score_value",
          "wdl_wins",
          "wdl_draws",
          "wdl_losses",
          "pv_uci",
          "depth"
        ],
        "title": "AnalysisObservationLineResponse",
        "type": "object"
      },
      "AnalysisObservationResponse": {
        "additionalProperties": false,
        "properties": {
          "fen": {
            "title": "Fen",
            "type": "string"
          },
          "result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AnalysisObservationResultResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "state": {
            "enum": [
              "not_requested",
              "queued",
              "running",
              "ready"
            ],
            "title": "State",
            "type": "string"
          }
        },
        "required": [
          "fen",
          "state",
          "result"
        ],
        "title": "AnalysisObservationResponse",
        "type": "object"
      },
      "AnalysisObservationResultResponse": {
        "additionalProperties": false,
        "properties": {
          "configuration_version": {
            "minimum": 1.0,
            "title": "Configuration Version",
            "type": "integer"
          },
          "engine_name": {
            "title": "Engine Name",
            "type": "string"
          },
          "engine_version": {
            "title": "Engine Version",
            "type": "string"
          },
          "lines": {
            "items": {
              "$ref": "#/components/schemas/AnalysisObservationLineResponse"
            },
            "title": "Lines",
            "type": "array"
          },
          "quality": {
            "enum": [
              "browser",
              "tool"
            ],
            "title": "Quality",
            "type": "string"
          },
          "settings": {
            "title": "Settings",
            "type": "object"
          },
          "terminal_kind": {
            "anyOf": [
              {
                "enum": [
                  "checkmate",
                  "stalemate",
                  "insufficient_material"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Terminal Kind"
          }
        },
        "required": [
          "quality",
          "configuration_version",
          "settings",
          "engine_name",
          "engine_version",
          "terminal_kind",
          "lines"
        ],
        "title": "AnalysisObservationResultResponse",
        "type": "object"
      },
      "AnalysisRequestBody": {
        "description": "Strict known request fields while tolerating future client fields.",
        "properties": {
          "fen": {
            "title": "Fen",
            "type": "string"
          },
          "quality": {
            "default": "browser",
            "enum": [
              "browser",
              "tool"
            ],
            "title": "Quality",
            "type": "string"
          }
        },
        "required": [
          "fen"
        ],
        "title": "AnalysisRequestBody",
        "type": "object"
      },
      "AnalysisRequestErrorResponse": {
        "additionalProperties": false,
        "description": "Stable error body for the desired-analysis operation.",
        "properties": {
          "code": {
            "enum": [
              "invalid_fen",
              "invalid_quality",
              "analysis_unavailable",
              "unexpected_failure"
            ],
            "title": "Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "AnalysisRequestErrorResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Chess Move Trainer",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/analysis": {
      "get": {
        "operationId": "getAnalysis",
        "parameters": [
          {
            "in": "query",
            "name": "fen",
            "required": true,
            "schema": {
              "title": "Fen",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisObservationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisObservationErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisObservationErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisObservationErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Get Analysis",
        "tags": [
          "analysis"
        ]
      }
    },
    "/api/analysis-requests": {
      "post": {
        "operationId": "requestAnalysis",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalysisRequestBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisObservationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisObservationResponse"
                }
              }
            },
            "description": "Accepted"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisRequestErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisRequestErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisRequestErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Post Analysis Request",
        "tags": [
          "analysis"
        ]
      }
    },
...
  }
}